+++ lib/trenino/application.ex
@@ -4,9 +4,12 @@ defmodule Trenino.Application do
   @moduledoc false
 
   use Application
+  require Logger
 
   @impl true
   def start(_type, _args) do
+    Logger.add_handlers(:trenino)
+
     children =
       [
         TreninoWeb.Telemetry,