+++ lib/trenino_web.ex
@@ -51,6 +51,7 @@ defmodule TreninoWeb do
   def live_view do
     quote do
       use Phoenix.LiveView
+      on_mount Sentry.LiveViewHook
 
       unquote(html_helpers())
     end